+++ src/main/java/de/uni_jena/cs/fusion/abecto/processor/UsePresentMappingProcessor.java
@@ -15,30 +15,20 @@
  */
 package de.uni_jena.cs.fusion.abecto.processor;
 
-import java.util.Collection;
+import java.util.Collections;
+import java.util.Map.Entry;
 import java.util.Objects;
 import java.util.Set;
 
-import org.apache.jena.datatypes.DatatypeFormatException;
-import org.apache.jena.query.Query;
-import org.apache.jena.query.QueryExecutionFactory;
-import org.apache.jena.query.QuerySolution;
-import org.apache.jena.query.ResultSet;
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.RDFNode;
 import org.apache.jena.rdf.model.Resource;
-import org.apache.jena.rdf.model.ResourceRequiredException;
-import org.apache.jena.sparql.core.TriplePath;
-import org.apache.jena.sparql.core.Var;
-import org.apache.jena.sparql.path.Path;
-import org.apache.jena.sparql.syntax.ElementPathBlock;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import de.uni_jena.cs.fusion.abecto.Aspect;
 import de.uni_jena.cs.fusion.abecto.Metadata;
 import de.uni_jena.cs.fusion.abecto.Parameter;
-import de.uni_jena.cs.fusion.abecto.datatype.SparqlPropertyPathType;
 
 public class UsePresentMappingProcessor extends MappingProcessor<UsePresentMappingProcessor> {
 	final static Logger log = LoggerFactory.getLogger(UsePresentMappingProcessor.class);
@@ -46,71 +36,50 @@ public class UsePresentMappingProcessor extends MappingProcessor<UsePresentMappi
 	@Parameter
 	public Resource aspect;
 	@Parameter
-	public Collection<String> assignmentPaths;
+	public String variable;
 
 	@Override
 	public void run() {
-
 		Model outputMetaModel = this.getOutputMetaModel(null);
 		Aspect aspectObject = Objects.requireNonNull(getAspects().get(this.aspect), "Unknown aspect.");
 
-		for (String unparsedAssignmentPath : this.assignmentPaths) {
+		// execute query for each dataset
+		for (Resource dataset : this.getDatasets()) {
+			// check if aspect pattern contains the variable
 			try {
-				// get path
-				Path assignmentPath = new SparqlPropertyPathType().parse(unparsedAssignmentPath);
-
-				// create query
-				Query query = new Query();
-				query.setQuerySelectType();
-				Var subject = Var.alloc("s");
-				Var object = Var.alloc("o");
-				query.addResultVar(subject);
-				query.addResultVar(object);
-				ElementPathBlock block = new ElementPathBlock();
-				block.addTriple(new TriplePath(subject, assignmentPath, object));
-				query.setQueryPattern(block);
-
-				// execute query for each dataset
-				for (Resource dataset : this.getDatasets()) {
-					Model inputPrimaryModel = this.getInputPrimaryModelUnion(dataset);
-
-					// get aspects resource
-					Set<Resource> aspectResources = Aspect.getResourceKeys(aspectObject, dataset, inputPrimaryModel);
-
-					ResultSet resultSet = QueryExecutionFactory.create(query, inputPrimaryModel).execSelect();
-					while (resultSet.hasNext()) {
-						QuerySolution solution = resultSet.next();
-						RDFNode node1 = solution.get("s");
-						RDFNode node2 = solution.get("o");
-						if (node1.isResource() && aspectResources.contains(node1.asResource())) {
-							try {
-								addCorrespondence(this.aspect, node1.asResource(), node2.asResource());
-							} catch (ResourceRequiredException e) {
-								Metadata.addIssue(node1.asResource(), null, node2, this.aspect, "Invalid Value",
-										String.format(
-												"Failed to get corresponding resource, found a literal: <%s> %s \"%s\"^^<%s>",
-												node1, assignmentPath, node2, node2.asLiteral().getDatatypeURI()),
-										outputMetaModel);
-							}
-						} else if (node2.isResource() && aspectResources.contains(node2.asResource())) {
-							try {
-								addCorrespondence(this.aspect, node1.asResource(), node2.asResource());
-							} catch (ResourceRequiredException e) {
-								Metadata.addIssue(node2.asResource(), null, node1, this.aspect, "Invalid Value",
-										String.format(
-												"Failed to get corresponding resource, found a literal: \"%s\"^^<%s> <%s> %s",
-												node1, node1.asLiteral().getDatatypeURI(), assignmentPath, node2),
-										outputMetaModel);
-							}
-						}
-						// else ignore
-					}
+				if (!aspectObject.getPattern(dataset).getResultVars().contains(variable)) {
+					log.warn("Missing variable(s) in pattern of aspect {} and dataset {}: {}", aspect, dataset,
+							variable);
+					return;
 				}
-			} catch (DatatypeFormatException e) {
-				throw new IllegalStateException("Failed to parse assignment path.", e);
+			} catch (NullPointerException e) {
+				log.warn("No pattern for aspect {} and dataset {} defined.", aspect, dataset);
+				return;
 			}
 
+			String varPath = aspectObject.getVarPathAsString(dataset, variable);
+
+			Model inputPrimaryModel = this.getInputPrimaryModelUnion(dataset);
+
+			for (Entry<RDFNode, Set<Resource>> entry : Aspect
+					.getResourceIndex(aspectObject, dataset, Collections.singletonList(variable), inputPrimaryModel)
+					.get(variable).entrySet()) {
+				RDFNode variableValue = entry.getKey();
+				Set<Resource> affectedResources = entry.getValue();
+				if (variableValue.isResource()) {
+					addCorrespondence(this.aspect, affectedResources, variableValue.asResource());
+				} else {
+					for (Resource affectedResource : entry.getValue())
+						Metadata.addIssue(affectedResource, variable, variableValue, this.aspect, "Invalid Value",
+								String.format(
+										"Failed to get corresponding resource, found a literal: <%s> %s \"%s\"^^<%s>",
+										affectedResource, varPath, variableValue,
+										variableValue.asLiteral().getDatatypeURI()),
+								outputMetaModel);
+				}
+			}
 		}
+
 		this.persistTransitiveCorrespondences();
 	}